refactor(gameplay): split clusterWeapons into focused modules
Extract hitTable, weaponSizes, resolution, streak from clusterWeapons.ts
into subdirectory.

- hitTable.ts: 210 lines - cluster hit table data and sizing
- weaponSizes.ts: 57 lines - weapon cluster size definitions
- resolution.ts: 146 lines - cluster hit rolling, location determination, attack resolution, grouping, formatting
- streak.ts: 61 lines - streak weapon detection and resolution
- index.ts: 36 lines - barrel export
- clusterWeapons.ts: 32 lines - facade re-exporting from subdirectory

All 42 cluster weapons tests pass. LSP clean. No breaking changes.
