+++ src/utils/gameplay/clusterWeapons/index.ts
@@ -1,29 +1,30 @@
 /**
  * Cluster Weapons Module
- * Facade re-exporting cluster weapon mechanics from subdirectory modules.
+ * Barrel export for cluster weapon mechanics.
  *
  * @spec openspec/changes/add-combat-resolution/specs/combat-resolution/spec.md
  */
 
+// Hit table and sizing
 export {
   CLUSTER_HIT_TABLE,
   CLUSTER_SIZES,
   type ClusterSize,
   getNearestClusterSize,
   lookupClusterHits,
-} from './clusterWeapons/hitTable';
+} from './hitTable';
 
-export {
-  CLUSTER_WEAPON_SIZES,
-  getClusterSizeForWeapon,
-} from './clusterWeapons/weaponSizes';
+// Weapon sizes
+export { CLUSTER_WEAPON_SIZES, getClusterSizeForWeapon } from './weaponSizes';
 
+// Resolution
 export {
   rollClusterHits,
   determineClusterHitLocations,
   resolveClusterAttack,
   groupClusterHitsByLocation,
   formatClusterResult,
-} from './clusterWeapons/resolution';
+} from './resolution';
 
-export { isStreakWeapon, resolveStreakAttack } from './clusterWeapons/streak';
+// Streak weapons
+export { isStreakWeapon, resolveStreakAttack } from './streak';
